Varint decoding is complete
Forgot to commit this earlier when it was completed.
Python handles integer storage with a varialbe number of bits so the
sign bit doesn't exist in the traditional sense.  Because of this,
signing consists of testing the traditional sign bit of the number
(32nd) and 2^33 (1 << 32) from the number if it should be negative.
This correctly negates the number in a way that python can interpret
natively.